requirements.txt
python-dotenv
gpiozero
requests
